[refactor] Place 내 websiteUrl 필드 추가 (#109)
