+++ src/main/java/org/recordy/server/place/domain/usecase/PlaceCreate.java
@@ -4,6 +4,7 @@ import org.recordy.server.location.domain.Location;
 
 public record PlaceCreate(
         String name,
+        String websiteUrl,
         Location location
 ) {
 }
