+++ src/test/java/org/recordy/server/util/PlaceFixture.java
@@ -7,16 +7,17 @@ import org.recordy.server.place.domain.usecase.PlaceCreate;
 public class PlaceFixture {
 
     public final static String NAME = "Place";
+    public final static String WEBSITE_URL = "https://www.google.com";
 
     public static Place create() {
-        return Place.create(new PlaceCreate(NAME, LocationFixture.create()));
+        return Place.create(new PlaceCreate(NAME, WEBSITE_URL, LocationFixture.create()));
     }
 
     public static Place create(Location location) {
-        return Place.create(new PlaceCreate(NAME, location));
+        return Place.create(new PlaceCreate(NAME, WEBSITE_URL, location));
     }
 
     public static Place create(String name, Location location) {
-        return Place.create(new PlaceCreate(name, location));
+        return Place.create(new PlaceCreate(name, WEBSITE_URL, location));
     }
 }
